main/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
opentelemetry-api==1.25.0 
opentelemetry-sdk==1.25.0 
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-exporter-otlp==1.25.0
opentelemetry.instrumentation.httpx==0.46b0